# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "暗水印清洗台后端",
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/detect": {
      "post": {
        "summary": "Detect",
        "operationId": "detect_api_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Detect Api Detect Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/collusion": {
      "post": {
        "summary": "Collusion",
        "description": "共谋平均：同一内容的多份不同水印副本对齐后平均（研究/授权测试）。",
        "operationId": "collusion_api_collusion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollusionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Collusion Api Collusion Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/similarity": {
      "post": {
        "summary": "Similarity",
        "operationId": "similarity_api_similarity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Similarity Api Similarity Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/scan": {
      "post": {
        "summary": "Import Scan",
        "operationId": "import_scan_api_import_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Scan Api Import Scan Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/frame": {
      "get": {
        "summary": "Frame",
        "description": "抽取视频指定时间点的帧（PNG），供预览与前后对比渲染。",
        "operationId": "frame_api_frame_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "default": 0.0,
              "title": "T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/thumb": {
      "get": {
        "summary": "Thumb",
        "description": "返回等比缩放的首帧缩略图（JPEG），带磁盘缓存，供素材封面使用。",
        "operationId": "thumb_api_thumb_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1920,
              "minimum": 16,
              "default": 320,
              "title": "Width"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library": {
      "get": {
        "summary": "Library List",
        "description": "返回持久化素材库（最新导入在前），并标记内容完全相同的重复副本。",
        "operationId": "library_list_api_library_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Library List Api Library Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Library Add",
        "description": "登记一个本地视频到素材库；内容完全相同的文件不重复导入。",
        "operationId": "library_add_api_library_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Library Add Api Library Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Library Remove",
        "description": "从素材库移除记录与对应产物；源视频文件一律保留，由用户手动管理。",
        "operationId": "library_remove_api_library_delete",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Library Remove Api Library Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media": {
      "get": {
        "summary": "Media",
        "description": "流式返回视频文件，支持 HTTP Range，播放器可随时拖动定位。",
        "operationId": "media_api_media_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollusionRequest": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 2,
            "title": "Paths"
          },
          "output": {
            "type": "string",
            "title": "Output"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mean",
              "median"
            ],
            "title": "Mode",
            "default": "mean"
          },
          "max_frames": {
            "type": "integer",
            "maximum": 6000.0,
            "minimum": 2.0,
            "title": "Max Frames",
            "default": 600
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paths",
          "output"
        ],
        "title": "CollusionRequest",
        "description": "共谋平均请求：同一内容的多份不同水印副本。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairRequest": {
        "properties": {
          "a": {
            "type": "string",
            "title": "A"
          },
          "b": {
            "type": "string",
            "title": "B"
          }
        },
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "title": "PairRequest"
      },
      "PathRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "PathRequest"
      },
...
      "ScanRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ScanRequest"
      },
...
    }
  }
}